// frontend/types/chat.ts
export interface UserContentItem {
  type: string;
  [key: string]: any; // This allows for additional properties based on the type
}

// Define the message format for object input
export interface UserMessageObject {
  role: 'user';
  content: UserContentItem[];
}

// Union type for the content parameter
export type UserChatContent = string | UserMessageObject;
// ...
export type ChatHistoryResponse = IChatDialogueMessageSchema[];
// ...
export type IChatDialogueMessageSchema = {
  role: 'human' | 'view' | 'system' | 'ai';
  context: string;
  order: number;
  time_stamp: number | string | null;
  model_name: string;
  retry?: boolean;
  thinking?: boolean;
  outing?: boolean;
  feedback?: Record<string, any>;
};